refac: 에러 핸들링 로직 분리
