+++ src/main/java/com/project/animal/missing/controller/MissingControllerAdvice.java
@@ -1,0 +1,90 @@
+package com.project.animal.missing.controller;
+
+import com.project.animal.global.common.utils.BindingResultParser;
+import com.project.animal.missing.constant.EndPoint;
+import com.project.animal.missing.constant.ViewName;
+import com.project.animal.missing.exception.*;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+
+import java.util.Map;
+
+@Slf4j
+@ControllerAdvice(basePackageClasses = MissingController.class)
+public class MissingControllerAdvice {
+
+
+  private final int SUCCESS_FLAG = 1;
+
+  private final int FAIL_FLAG = 0;
+
+  @ExceptionHandler(DetailNotFoundException.class)
+  public String handleDetailNotFound(DetailNotFoundException ex, Model model) {
+    model.addAttribute("error", "Fail to find detail");
+    model.addAttribute("type", "detail");
+
+    return ViewName.POST_DETAIL;
+  }
+
+  @ExceptionHandler(InvalidCreateFormException.class)
+  public String handleInvalidCreateForm(InvalidCreateFormException ex, RedirectAttributes redirectAttributes) {
+
+    Map<String, String> errors = BindingResultParser.parse(ex.getBindingResult());
+
+    redirectAttributes.addFlashAttribute("isRedirected", SUCCESS_FLAG);
+    redirectAttributes.addFlashAttribute("isSuccess", FAIL_FLAG);
+    redirectAttributes.addFlashAttribute("errors", errors);
+    redirectAttributes.addFlashAttribute("msg", "입력한 정보를 다시 확인해주세요.");
+    redirectAttributes.addFlashAttribute("post", ex.getInvalidForm());
+
+    log.error("handleInvalidCreateForm: >> Invalid Input " + errors.toString());
+    return "redirect:" + EndPoint.MISSING + EndPoint.NEW;
+  }
+
+  @ExceptionHandler(PostSaveFailException.class)
+  public String handlePostSaveFail(PostSaveFailException ex, RedirectAttributes redirectAttributes) {
+
+    redirectAttributes.addFlashAttribute("isRedirected", SUCCESS_FLAG);
+    redirectAttributes.addFlashAttribute("isSuccess", FAIL_FLAG);
+    redirectAttributes.addFlashAttribute("errors", ex.getMessage());
+    redirectAttributes.addFlashAttribute("msg", "입력한 정보를 다시 확인해주세요.");
+    redirectAttributes.addFlashAttribute("post", ex.getInvalidForm());
+
+    log.error("handlePostSaveFail: >> Invalid Input " + ex.getMessage());
+    return "redirect:" + EndPoint.MISSING + EndPoint.NEW;
+  }
+
+  @ExceptionHandler(PostDeleteFailException.class)
+  public String handlePostDeleteFail(PostDeleteFailException ex, RedirectAttributes redirectAttributes) {
+    redirectAttributes.addFlashAttribute("error", "Fail to delete post");
+    redirectAttributes.addFlashAttribute("type", "delete");
+
+    return "redirect:" + EndPoint.MISSING + "/" + ex.getTargetId();
+  }
+
+  @ExceptionHandler(PostEditFailException.class)
+  public String handlePostEditFail(PostEditFailException ex, RedirectAttributes redirectAttributes) {
+    redirectAttributes.addFlashAttribute("error", "Fail to edit post");
+    redirectAttributes.addFlashAttribute("type", "edit");
+
+    return "redirect:" + EndPoint.MISSING + "/" + ex.getId();
+  }
+
+  @ExceptionHandler(InvalidEditFormException.class)
+  public String handleInvalidEditFormException(InvalidEditFormException ex, RedirectAttributes redirectAttributes) {
+    Map<String, String> errors = BindingResultParser.parse(ex.getBindingResult());
+
+    redirectAttributes.addFlashAttribute("isRedirected", SUCCESS_FLAG);
+    redirectAttributes.addFlashAttribute("isSuccess", FAIL_FLAG);
+    redirectAttributes.addFlashAttribute("errors", errors);
+    redirectAttributes.addFlashAttribute("msg", "입력한 정보를 다시 확인해주세요.");
+    redirectAttributes.addFlashAttribute("post", ex.getInvalidForm());
+
+    log.error("InvalidEditFormException: >> Invalid Input " + errors.toString());
+    return "redirect:" + EndPoint.MISSING + "/edit" + '/' + ex.getInvalidForm().getMissingId();
+  }
+
+}
